+++ pysgrid/read_netcdf.py
@@ -3,8 +3,53 @@ Created on Mar 19, 2015
 
 @author: ayan
 '''
+import re
+
 from .lookup import LAT_GRID_CELL_NODE_LONG_NAME, LON_GRID_CELL_NODE_LONG_NAME
-from .utils import parse_padding
+from .utils import GridPadding
+from .custom_exceptions import CannotFindPaddingError
+
+
+def parse_padding(padding_str, mesh_topology_var):
+    """
+    Use regex expressions to break apart an
+    attribute string containing padding types
+    for each variable with a cf_role of 
+    'grid_topology'.
+    
+    Padding information is returned within a named tuple
+    for each node dimension of an edge, face, or vertical
+    dimension. The named tuples have the following attributes:
+    mesh_topology_var, dim_name, dim_var, and padding.
+    Padding information is returned as a list
+    of these named tuples.
+    
+    :param str padding_str: string containing padding types from a netCDF attribute
+    :return: named tuples with padding information
+    :rtype: list
+    
+    """
+    p = re.compile('([a-zA-Z0-9_]+:) ([a-zA-Z0-9_]+) (\(padding: [a-zA-Z]+\))')
+    padding_matches = p.findall(padding_str)
+    padding_type_list = []
+    for padding_match in padding_matches:
+        raw_dim, raw_sub_dim, raw_padding_var = padding_match
+        dim = raw_dim.split(':')[0]
+        sub_dim = raw_sub_dim
+        cleaned_padding_var = re.sub('[\(\)]', '', raw_padding_var)  # remove parentheses
+        padding_type = cleaned_padding_var.split(':')[1].strip()  # get the padding value and remove spaces
+        grid_padding = GridPadding(mesh_topology_var=mesh_topology_var,
+                                   dim=dim,
+                                   sub_dim=sub_dim,
+                                   padding=padding_type
+                                   )
+        padding_type_list.append(grid_padding)
+    if len(padding_type_list) > 0:
+        final_padding_types = padding_type_list
+    else:
+        final_padding_types = None
+        raise CannotFindPaddingError
+    return final_padding_types  
 
 
 def make_coordinate_parser(attr_name, find_coordinates_manually):
