+++ pysgrid/tests/test_utils.py
@@ -5,61 +5,8 @@ Created on Mar 23, 2015
 '''
 import unittest
 import numpy as np
-from ..utils import parse_padding, pair_arrays, check_element_equal
-from ..custom_exceptions import CannotFindPaddingError
+from ..utils import pair_arrays, check_element_equal
 
-
-class TestParsePadding(unittest.TestCase):
-    
-    def setUp(self):
-        self.grid_topology = 'some_grid'
-        self.with_two_padding = 'xi_rho: xi_psi (padding: both) eta_rho: eta_psi (padding: low)'
-        self.with_one_padding = 'xi_v: xi_psi (padding: high) eta_v: eta_psi'
-        self.with_no_padding = 'MMAXZ: MMAX NMAXZ: NMAX'
-        
-    def test_mesh_name(self):
-        result = parse_padding(self.with_one_padding, self.grid_topology)
-        mesh_topology = result[0].mesh_topology_var
-        expected = 'some_grid'
-        self.assertEqual(mesh_topology, expected)
-    
-    def test_two_padding_types(self):
-        result = parse_padding(self.with_two_padding, self.grid_topology)
-        expected_len = 2
-        padding_datum_0 = result[0]
-        padding_type = padding_datum_0.padding
-        sub_dim = padding_datum_0.sub_dim
-        dim = padding_datum_0.dim
-        expected_sub_dim = 'xi_psi'
-        expected_padding_type = 'both'
-        expected_dim = 'xi_rho'
-        self.assertEqual(len(result), expected_len)
-        self.assertEqual(padding_type, expected_padding_type)
-        self.assertEqual(sub_dim, expected_sub_dim)
-        self.assertEqual(dim, expected_dim)
-        
-    def test_one_padding_type(self):
-        result = parse_padding(self.with_one_padding, self.grid_topology)
-        expected_len = 1
-        padding_datum_0 = result[0]
-        padding_type = padding_datum_0.padding
-        sub_dim = padding_datum_0.sub_dim
-        dim = padding_datum_0.dim
-        expected_padding_type = 'high'
-        expected_sub_dim = 'xi_psi'
-        expected_dim = 'xi_v'
-        self.assertEqual(len(result), expected_len)
-        self.assertEqual(padding_type, expected_padding_type)
-        self.assertEqual(sub_dim, expected_sub_dim)
-        self.assertEqual(dim, expected_dim)
-        
-    def test_no_padding(self):
-        self.assertRaises(CannotFindPaddingError, 
-                          parse_padding, 
-                          padding_str=self.with_no_padding,
-                          mesh_topology_var=self.grid_topology
-                          )
-        
 
 class TestPairArrays(unittest.TestCase):
